requirements.txt
python-dotenv
Flask
pandas
numpy
scikit-learn
PyGithub
gunicorn